# tokpulse-railway-fixed/tokpulse-build/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TokPulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tiktok/connect:
    post:
      operationId: connectTikTok
      tags: [tiktok]
      summary: Connect a TikTok account via OAuth code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TikTokConnectInput"
      responses:
        "200":
          description: TikTok connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TikTokAccount"
...
  /tiktok/account:
    get:
      operationId: getTikTokAccount
      tags: [tiktok]
      summary: Get connected TikTok account info
      responses:
        "200":
          description: TikTok account info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TikTokAccount"
        "404":
          description: No TikTok account connected
...
components:
  schemas:
...
    TikTokAccount:
      type: object
      required: [id, username, displayName, followerCount, followingCount, videoCount]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        followerCount:
          type: integer
        followingCount:
          type: integer
        videoCount:
          type: integer
        lastSyncedAt:
          type: ["string", "null"]
    TikTokConnectInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
        redirectUri:
          type: string
...